pyproject.toml
[tool.poetry]
name = "doc-modifier"
version = "0.1.0"
description = "test streamlit project"
authors = ["vodenikov.daniil@yandex.ru"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
streamlit = ">=1.45.1,<2.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
python-docx = ">=1.1.2,<2.0.0"
loguru = "^0.7.3"
langgraph = "^0.4.8"
langchain = "^0.3.25"
langchain-google-genai = "^2.1.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
